Replace bat commands with cat in various scripts
The commit replaces bat commands with cat in several scripts to fetch and display data from file. This change potentially improves the compatibility of the scripts with environments where 'bat' command may not be available. Also, a phrase in the penz.sh script is corrected, and a requirement to have 'fd' installed is added in README.md.